requirements.txt
astroid==2.1.0
awsebcli==3.17.1
backports.csv==1.0.7
botocore==1.14.13
cement==2.8.2
certifi==2019.3.9
chardet==3.0.4
colorama==0.3.9
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.5
django-cors-headers==2.1.0
django-import-export==1.2.0
django-uuidfield==0.5.0
djangorestframework==3.10.3
djangorestframework-simplejwt==3.3
docutils==0.15.2
et-xmlfile==1.0.1
fcm-django==0.2.21
future==0.16.0
geographiclib==1.49
geopy==1.19.0
idna==2.7
isort==4.2.15
jdcal==1.4.1
jmespath==0.9.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
odfpy==1.4.0
openpyxl==2.6.2
pathspec==0.5.9
psycopg2==2.8.3
psycopg2-binary==2.8.2
pyfcm==1.4.5
PyJWT==1.7.1
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
pypiwin32==223
pyshp==2.1.0
python-dateutil==2.8.0
pytz==2017.2
pywin32==227
PyYAML==5.2
requests==2.20.1
requests-toolbelt==0.9.1
semantic-version==2.5.0
Shapely==1.6.4.post2
six==1.11.0
sqlparse==0.3.0
tablib==0.13.0
termcolor==1.1.0
typed-ast==1.1.1
urllib3==1.24.3
wcwidth==0.1.8
wrapt==1.10.11
xlrd==1.2.0
xlwt==1.3.0
